+++ src/components/Treatment/index.ts
@@ -1,1 +1,2 @@
-export { TreatmentFunction, Treatment, TreatmentVariant } from "./Treatment";
+export { Treatment, TreatmentVariant } from "./Treatment";
+export { TreatmentFunction } from "./TreatmentFunction";
