+++ src/index.ts
@@ -1,26 +1,24 @@
+import { SDKProvider } from "./components/SDKProvider";
 import {
-  SDKProvider,
-  useABSmartly,
-  withABSmartly,
-} from "./components/SDKProvider";
-import {
-  TreatmentFunction,
   Treatment,
+  TreatmentFunction,
   TreatmentVariant,
 } from "./components/Treatment";
 import { useTreatment } from "./hooks/useTreatment";
+
+import { withABSmartly } from "./hooks/HOCs/withABSmartly";
+import { useABSmartly } from "./hooks/useABSmartly";
+
 import { mergeConfig } from "@absmartly/javascript-sdk";
 import {
   ABSmartly,
-  TreatmentProps,
-  ABSmartlySDK,
   ABSmartlyContext,
+  ABSmartlySDK,
+  TreatmentProps,
 } from "./types";
 
 export * from "@absmartly/javascript-sdk";
-
 export type { ABSmartly, TreatmentProps, ABSmartlySDK, ABSmartlyContext };
-
 export {
   mergeConfig,
   Treatment,
